feat(task): add category feature
configure local development and apply UI and fetch data from localhost

closes #23

+++ server/nhost/migrations/default/1658426144792_create_table_public_categories/up.sql
@@ -5,7 +5,7 @@ CREATE TABLE "public"."categories" (
     "user_id" uuid NOT NULL,
     "created_at" timestamptz NOT NULL DEFAULT NOW(),
     "update_at" timestamptz,
-    "color" int4,
+    "color" BIGINT,
     PRIMARY KEY ("id"),
     FOREIGN KEY ("user_id") REFERENCES "auth"."users" ("id") ON UPDATE RESTRICT ON DELETE CASCADE
 );
